Move dataframe logic from tileClient to MVT
