+++ src/sources/DataframeCache.js
@@ -18,20 +18,15 @@ export default class DataframeCache {
         this._cache = LRU(lruOptions);
     }
 
-    // Get the promise of the dataframe with the provided unique ID, by querying the local cache, and using the fetch function as a fallback.
-    // The `fetch` function will be called with the provided `uid` and it is expected that it will return a promise to a Dataframe
-    get(uid, fetch) {
-        const cachedDataframe = this._cache.get(uid);
-        if (cachedDataframe) {
-            return cachedDataframe;
-        }
-
-        const dataframePromise = fetch(uid);
-        this._cache.set(uid, dataframePromise);
-        return dataframePromise;
+    get(key) {
+        return this._cache.get(key);
     }
 
-    free(){
+    set(key, value) {
+        return this._cache.set(key, value);
+    }
+
+    free() {
         this._cache.reset();
     }
 }
